Fix qorca42 to work with PALn and multiple jobs
Problem: qorca42 searches over the entire file for lines beginning with
!, and the last match is taken to be keyword_line.
Solution: Break out of the loop once a line beginning with ! is found.
This ensures that the first line is used. This avoids issues with
detection of the PALn keyword (which only needs to be specified in the
keywords of the first job).